Replace the node-based REPL with an MSBuild-based one
The advantage is that it should work on anyone's machine without any additional dependencies:
- You no longer need `node`. Instead, you need MSBuild and PowerShell which everybody should have.
- When building `EvalProgram.cs`, the reference paths are received from MSBuild rather than hard coded in a script. Consequently, the REPL build process won't break if your Beat Saber or .NETFramework directories are at different paths than mine.

The disadvantages are that it's slower and more complicated. Evaluating an expression takes a couple of seconds instead of under a second. Most of the time is probably spent waiting on MSBuild to do its work.