Fix compatability with CustomJSONData
Infinite mode wasn't working when the CustomJSONData plugin was installed.

When starting infinite mode, Infinite Beat Saber would call `BeatmapData.InsertBeatmapEventDataInOrder` which would lead to the exception:

```
KeyNotFoundException: The given key 'CustomJSONData.CustomBeatmap.CustomBPMChangeBeatmapEventData' was not present in the dictionary.
```

The problem was that CustomJSONData converted all of the map's `BeatmapDataItems` to its own `BeatmapDataItem` subclasses (e.g. `CustomBPMChangeBeatmapEventData`). Part of Beat Saber's `InsertBeatmapEventDataInOrder` implementation blows up when given a `BeatmapDataItem` with which it isn't familiar.

`BeatmapRemixer.AddCustomJSONDataProcessors` fixes this problem. See that method's comment for details.

The solution required Infinite Beat Saber to depend on CustomJSONData and depending on CustomJSONData required updating Infinite Beat Saber from .NET Framework 4.7.2 to 4.8.