Add REPL tests for song catalog capatibility
Added a test method that can be run from the REPL called `GenerateSongCatalogCompatibilityReport`. It iterates over the Beat Saber song catalog and reports any things that could make a level incompatible with Infinite Beat Saber.

I also commented out an assert that fails for some songs in `InfiniteRemix.cs`. The check is now included in the REPL test `GenerateSongCatalogCompatibilityReport`. One reason the assert is failing is that, for some songs, the BPM determined by Spotify is a multiple of the BPM in the Beat Saber data file (e.g. Beat Saber BPM is 80, Spotify BPM is 160). The assert uses the Beat Saber BPM to evaluate Spotify beats so it falsely fails when the Beat Saber & Spotify BPMs don't match.

Anyway, it's better to check and catch these issues at test time rather than runtime.