Trim end of Spotify audio analysis data so it can't be longer than the Beat Saber song
Some Spotify songs are longer than their Beat Saber counterparts.

Currently, if the Spotify audio analysis data has timestamps that are outside of the duration of the Beat Saber song, then the remix could include these timestamps that are out of bounds. The result is that the player hears silence.

This happens with Ghost. When the remix reaches the section at the end of the song that only has 1 branch coming out of it (that leads back to the beginning of the song) the player hears silence during that section because it's outside of the duration of the Beat Saber song. The overall feel is that the song ends and then starts again (as though your music player was configured for repeat) breaking the infinite song experience.

This commit fixes the issue by removing anything from the Spotify analysis that has a timestamp that is larger than the duration of the Beat Saber song.