Add support for `LightColorBeatmapEventData` & `LightRotationBeatmapEventData`
This implementation seems to work fine but I suspect it could be better.

I suspect that the lights can look differently in a section depending on what sections were played before it. So when playing a section in Infinite Beat Saber mode, the lights may not look as intended by the level designer.

This implementation seems good enough for now. We should investigate whether a better implementation is needed.