Support Beat Saber 1.29.1 (in addition to 1.31.1)
The main challenge was that 2 things needed to be different based on the version of Beat Saber we're building for:
- `manifest.json`
  - `gameVersion` needs to exactly match the version of Beat Saber we're targeting (1.29.1 or 1.31.1).
  - `dependsOn`. We might want to specify different versions of dependencies based on the version of Beat Saber we're targeting.
- `MediaAsyncLoaderPatches` needs conditional compilation so it can change 1 line of code based on the Beat Saber version due to a breaking change in Unity's `UnityWebRequest` API.

The solution involved introducing new Visual Studio Configurations so we could vary the Beat Saber verison in addition to Debug/Release. So we now have:
- Beat_Saber_1.29.1_Debug
- Beat_Saber_1.29.1_Release
- Debug
- Release

These vary these new MSBuild properties:
- BuildType (Debug or Release)
- BeatSaberVersion (1.29.1 or Latest)

To solve the manifest problem, I upgarded to a newer version of BeatSaberModdingTools.Tasks which supports specifying the manifest using MSBuild properties (so we can leverage MSBuild's conditional construct). Setting the MSBuild property `GenerateManifest` to `true` enables this feature. We specify the parts of the manifest that need to vary by Beat Saber version ("gameVersion" and "dependsOn") as MSBuild properties in the csproj file.

Originally, I tried to put all of the manifest stuff into a "Manifest.props" file and import that into the csproj but there seemed to be an evaluation bug when changing configurations. When switching from Beat_Saber_1.29.1_Debug to Debug, the build would fail and it seemed that "DependsOn" had no value. Making a no-op edit to the csproj file so VS would reload it seemed to fix this. That's annoying so I decided to abandon the "Manifest.props" file in favor of specifying the properties directly in the csproj file (which didn't exhibit this bug).

I updated my csproj.user file to use different Beat Saber paths based on the Configuration:

```
<PropertyGroup Condition="'$(BeatSaberVersion)' == '1.29.1'">
  <BeatSaberDir>C:\Program Files (x86)\Steam\steamapps\common\Beat Saber (1.29.1)</BeatSaberDir>
</PropertyGroup>
<PropertyGroup Condition="'$(BeatSaberVersion)' == 'Latest'">
  <BeatSaberDir>C:\Program Files (x86)\Steam\steamapps\common\Beat Saber (1.31.1)</BeatSaberDir>
</PropertyGroup>
```

When the MSBuild property `BeatSaberVersion` is `1.29.1`, the constant `BEAT_SABER_1_29_1` is set and can be checked in C# using `#if`. This enabled us to use conditional compilation to mitigate the breaking change in `UnityWebRequest`.

Note that the REPL technique (implemented by the `Eval` project) is severely limited in Beat Saber 1.29.1. For reasons I don't understand, only the first eval works properly. Subsequent evals rerun the code from the first eval rather than running the new code you're trying to eval.